+++ components/miot_th/miot_th.cpp
@@ -1,28 +1,28 @@
 #include "esphome/core/log.h"
-#include "miot_thermogigro.h"
+#include "miot_th.h"
 
 namespace esphome {
-namespace miot_thermogigro {
+namespace miot_th {
 
-static const char *const TAG = "miot_thermogigro";
+static const char *const TAG = "miot_th";
 
 constexpr uint16_t PRODUCT_ID_XMWSDJ04MMC = 0x1203;
 constexpr uint16_t PRODUCT_ID_LYWSD02MMC = 0x16E4;
 
-void MiotThermoGigro::dump_config() {
+void MiotTH::dump_config() {
   this->dump_config_(TAG);
   LOG_SENSOR("  ", "Temperature", this);
   LOG_SENSOR("  ", "Humidity", this->humidity_);
 }
 
-void MiotThermoGigro::process_temperature_(const miot::BLEObject &obj) {
+void MiotTH::process_temperature_(const miot::BLEObject &obj) {
   const auto temperature = obj.get_temperature();
   if (temperature.has_value()) {
     this->publish_state(*temperature);
   }
 }
 
-void MiotThermoGigro::process_humidity_(const miot::BLEObject &obj) {
+void MiotTH::process_humidity_(const miot::BLEObject &obj) {
   if (this->humidity_ != nullptr) {
     const auto humidity = obj.get_humidity();
     if (humidity.has_value()) {
@@ -31,7 +31,7 @@ void MiotThermoGigro::process_humidity_(const miot::BLEObject &obj) {
   }
 }
 
-void MiotThermoGigro::process_temperature_humidity_(const miot::BLEObject &obj) {
+void MiotTH::process_temperature_humidity_(const miot::BLEObject &obj) {
   const auto temperature_humidity = obj.get_temperature_humidity();
   if (temperature_humidity != nullptr) {
     this->publish_state(temperature_humidity->get_temperature());
@@ -41,48 +41,39 @@ void MiotThermoGigro::process_temperature_humidity_(const miot::BLEObject &obj) 
   }
 }
 
-void MiotThermoGigro::process_miaomiaoce_temperature_(const miot::BLEObject &obj) {
-  // if (this->product_id_ != PRODUCT_ID_XMWSDJ04MMC) {
-  //   return;
-  // }
-  const auto temperature = obj.get_miaomiaoce_temperature();
+void MiotTH::process_miaomiaoce_temperature_1001_(const miot::BLEObject &obj) {
+  const auto temperature = obj.get_miaomiaoce_temperature_1001();
   if (temperature.has_value()) {
     this->publish_state(*temperature);
   }
 }
 
-void MiotThermoGigro::process_miaomiaoce_humidity_(const miot::BLEObject &obj) {
-  // if (this->product_id_ != PRODUCT_ID_XMWSDJ04MMC) {
-  //   return;
-  // }
+void MiotTH::process_miaomiaoce_humidity_1008_(const miot::BLEObject &obj) {
   if (this->humidity_ != nullptr) {
-    const auto humidity = obj.get_miaomiaoce_humidity();
+    const auto humidity = obj.get_miaomiaoce_humidity_1008();
     if (humidity.has_value()) {
       this->humidity_->publish_state(*humidity);
     }
   }
 }
 
-void MiotThermoGigro::process_miaomiaoce_humidity_o2_(const miot::BLEObject &obj) {
-  // if (this->product_id_ != PRODUCT_ID_LYWSD02MMC) {
-  //   return;
-  // }
+void MiotTH::process_miaomiaoce_humidity_1002_(const miot::BLEObject &obj) {
   if (this->humidity_ != nullptr) {
-    const auto humidity = obj.get_miaomiaoce_sensor_ht_o2_humidity();
+    const auto humidity = obj.get_miaomiaoce_humidity_1002();
     if (humidity.has_value()) {
       this->humidity_->publish_state(*humidity);
     }
   }
 }
 
-bool MiotThermoGigro::process_mibeacon(const miot::MiBeacon &mib) {
+bool MiotTH::process_mibeacon(const miot::MiBeacon &mib) {
   if (this->product_id_ == 0) {
     this->product_id_ = mib.product_id;
   }
   return MiotComponent::process_mibeacon(mib);
 }
 
-bool MiotThermoGigro::process_object_(const miot::BLEObject &obj) {
+bool MiotTH::process_object_(const miot::BLEObject &obj) {
   switch (obj.id) {
     case miot::MIID_TEMPERATURE:
       this->process_temperature_(obj);
@@ -96,16 +87,16 @@ bool MiotThermoGigro::process_object_(const miot::BLEObject &obj) {
       this->process_temperature_humidity_(obj);
       break;
 
-    case miot::MIID_MIAOMIAOCE_TEMPERATURE:
-      this->process_miaomiaoce_temperature_(obj);
+    case miot::MIID_MIAOMIAOCE_TEMPERATURE_1001:
+      this->process_miaomiaoce_temperature_1001_(obj);
       break;
 
-    case miot::MIID_MIAOMIAOCE_HUMIDITY:
-      this->process_miaomiaoce_humidity_(obj);
+    case miot::MIID_MIAOMIAOCE_HUMIDITY_1008:
+      this->process_miaomiaoce_humidity_1008_(obj);
       break;
 
-    case miot::MIID_MIAOMIAOCE_SENSOR_HT_O2_HUMIDITY:
-      this->process_miaomiaoce_humidity_o2_(obj);
+    case miot::MIID_MIAOMIAOCE_HUMIDITY_1002:
+      this->process_miaomiaoce_humidity_1002_(obj);
       break;
 
     default:
@@ -114,5 +105,5 @@ bool MiotThermoGigro::process_object_(const miot::BLEObject &obj) {
   return true;
 }
 
-}  // namespace miot_thermogigro
+}  // namespace miot_th
 }  // namespace esphome